fix: best practices and cleanups from automated review (#1115)
- Replace mutable default storage args in numpy.py with None + lazy init
- Fix AxesTuple.index return type from tuple[float,...] to tuple[int,...]
- Fix Self = object fallback to Self = typing.Any in _compat/typing.py
- Remove duplicate BHP_NOEXCEPT_17 macro in register_axis.hpp
- Guard division by zero in mean.hpp and weighted_mean.hpp variance()
- Remove leftover print() debug statements in test files
- Remove commented-out dead test_hist_division code
- Make TestBoolean inherit from Axis ABC for structural coverage
- Add hypothesis.settings(deadline=None) to property-based test files

+++ src/boost_histogram/axis/__init__.py
@@ -890,7 +890,7 @@ class AxesTuple(tuple):  # type: ignore[type-arg]
             )
         return tuple(self[i].bin(indexes[i]) for i in range(len(indexes)))
 
-    def index(self, *values: float) -> tuple[float, ...]:  # type: ignore[override, override]
+    def index(self, *values: float) -> tuple[int, ...]:  # type: ignore[override, override]
         if len(values) != len(self):
             raise IndexError(
                 "Must have the same number of arguments as the number of axes"
